backend/requirements.txt
fastapi>=0.115.0
uvicorn==0.24.0
mangum==0.17.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.10.0
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pyjwt>=2.8.0
supabase>=2.0.0